Refactors single route app config
