+++ config/extenders/codeSplitting.js
